+++ diive/pkgs/createvar/conversions.py
@@ -1,6 +1,26 @@
 import pandas as pd
 
 
+def air_temp_from_sonic_temp(sonic_temp: pd.Series, h2o: pd.Series) -> pd.Series:
+    """
+    Calculate air temperature from sonic temperature and water vapor concentration.
+
+    This function computes the air temperature from the provided sonic temperature
+    and water vapor concentration using a specific formula. The calculation takes
+    into account the relationship between sonic temperature, air temperature, and
+    the effect of water vapor on sonic wave speed.
+
+    Args:
+        sonic_temp (pd.Series): Sonic temperature data in Kelvin.
+        h2o (pd.Series): Water vapor concentration in the air in mol mol-1.
+
+    Returns:
+        pd.Series: Air temperature data in Kelvin.
+    """
+    ta = sonic_temp / (1 + 0.32 * h2o)
+    return ta
+
+
 def latent_heat_of_vaporization(ta: pd.Series) -> pd.Series:
     """Calculate latent heat of vaporization as a function of air temperature.